Fix sorting of point releases on rakudo.org/downloads/
PHP's glob() gives them to us AFTER non-point releases, so when a
point release is the latest release, we end up with non-point release
in the "latest" box.

Fix by shoving them into a separate array first and prepend to file
list when we get the first non-point release (or we're done looking
through files).